backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-groq
langchain-community
faiss-cpu
crewai
openai